source/applications/lib/fastapi-core/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
    # via fastapi-core (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
asgiref==3.10.0
    # via opentelemetry-instrumentation-asgi
charset-normalizer==3.4.4
    # via core
click==8.3.0
    # via uvicorn
deprecated==1.3.1
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-semantic-conventions
fastapi==0.115.14
    # via fastapi-core (pyproject.toml)
googleapis-common-protos==1.72.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via uvicorn
idna==3.11
    # via anyio
importlib-metadata==8.6.1
    # via opentelemetry-api
opentelemetry-api==1.32.1
    # via
    #   fastapi-core (pyproject.toml)
    #   core
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-distro==0.53b1
    # via fastapi-core (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.32.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.32.1
    # via
    #   fastapi-core (pyproject.toml)
    #   core
opentelemetry-instrumentation==0.53b1
    # via
    #   opentelemetry-distro
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-asgi==0.53b1
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.53b1
    # via fastapi-core (pyproject.toml)
opentelemetry-instrumentation-httpx==0.53b1
    # via core
opentelemetry-proto==1.32.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.32.1
    # via
    #   fastapi-core (pyproject.toml)
    #   core
    #   opentelemetry-distro
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.53b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-sdk
opentelemetry-util-http==0.53b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
packaging==25.0
    # via opentelemetry-instrumentation
prometheus-client==0.23.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.1.0
    # via fastapi-core (pyproject.toml)
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.11.10
    # via
    #   fastapi-core (pyproject.toml)
    #   core
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.34.3
    # via fastapi-core (pyproject.toml)
wrapt==1.17.3
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
zipp==3.23.0
    # via importlib-metadata
